Mostly revamps, some fixes
- Made BuffEffects a singleton
- Added weapon multiplier to stat calculations
- Revamped attack system to support delayed attacks
- 'zigzag' flag in animations is handled
- 'noflip' flag in mobs is handled
- Split UI into seperate 'Login' and 'Game' versions
- Changed some more enums to new format
- Removed some unneccessary files
- Moved some files from 'Program' and 'Util' to project folder
- Lots of renaming for shortening or improving names

+++ Character/Look/Body.cpp
@@ -20,7 +20,7 @@
 
 namespace Character
 {
-	Body::Body(uint16_t skin, const BodyDrawinfo& drawinfo)
+	Body::Body(int32_t skin, const BodyDrawinfo& drawinfo)
 	{
 		string sk;
 		if (skin < 10)
